Added English to README.md
Added English translation throughout the document.  I corrected grammar and sentence structure along-the-way (because I used Google Translate for the initial translation), but I always welcome feedback and corrections.

Hopefully this will help get more users interested in this GVFS add-on.  I'm certainly interested in it because I used the Netbox add-on multiple times daily when I was on Windows.  Now that I'm on Linux, I'm finding that I need a replacement since far2l does not have a built-in SFTP option (I'm very pleased with everything that has been done to port Far Manager to Linux, and my hope is that development continues -- thank you to all developers for all of your wonderful and excellent efforts, I appreciate it tremendously).